dashboards/plotly/stock_dashboard/requirements.txt
plotly
yfinance
dash-core-components
dash-html-components
dash-bootstrap-components
scipy
numpy
pandas
pymongo
python-dotenv
os
